# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Hootly Public API
  version: 0.6.0
  description: >
    Contrat public Hootly. Authentification (Sprint 02), profil et marques
    (Sprint 03), publications, médias, planification et calendrier (Sprint 04),
    comptes sociaux et OAuth Meta (Sprint 06/07), commentaires et réponses
    contrôlées (Sprint 08), analyse de sentiment (Sprint 09), propositions de
    réponse et hashtags (Sprint 10), notifications et Firebase Cloud
    Messaging (Sprint 11).
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/auth/register:
    post:
      summary: Créer un compte utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { $ref: '#/components/responses/Session' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/auth/login:
    post:
      summary: Ouvrir une session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/refresh:
    post:
      summary: Faire tourner le refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/auth/logout:
    post:
      summary: Révoquer la session courante
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Session révoquée }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/auth/me:
    get:
      summary: Retourner l'utilisateur et sa marque active
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/CurrentUser' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/auth/forgot-password:
    post:
      summary: Demander un lien de réinitialisation sans révéler l’existence du compte
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '202': { description: Demande acceptée, réponse identique pour toute adresse }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/reset-password:
    post:
      summary: Utiliser une fois un token de réinitialisation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '204': { description: Mot de passe réinitialisé et sessions révoquées }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/auth/change-password:
    post:
      summary: Modifier le mot de passe de l’utilisateur connecté
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/profile:
    get:
      summary: Lire le profil de l’utilisateur connectÃ©
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
    patch:
      summary: Modifier le profil de l’utilisateur connectÃ©
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProfilePatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/profile/preferences:
    get:
      summary: Lire les prÃ©fÃ©rences du profil
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
    patch:
      summary: Modifier les prÃ©fÃ©rences du profil
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PreferencesPatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/profile/avatar:
    post:
      summary: Associer un avatar prÃ©alablement dÃ©posÃ© au profil
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AvatarMetadata' } } }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      summary: Retirer l’avatar du profil
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Avatar retirÃ© } }
  /api/v1/brands:
    get:
      summary: Lister les marques auxquelles l’utilisateur appartient
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: active, in: query, schema: { type: boolean } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
    post:
      summary: CrÃ©er une marque et l’activer pour son propriÃ©taire
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BrandCreate' } } }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/brands/{brandId}:
    parameters:
      - { name: brandId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Lire une marque autorisÃ©e
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Modifier une marque (OWNER ou ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BrandPatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Archiver une marque (OWNER uniquement)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Marque archivÃ©e }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/brands/{brandId}/activate:
    post:
      summary: DÃ©finir la marque active de l’utilisateur connectÃ©
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: brandId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/brands/{brandId}/ai-settings:
    parameters:
      - { name: brandId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Lire les paramÃ¨tres IA de la marque
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
    patch:
      summary: CrÃ©er une nouvelle version des paramÃ¨tres IA (OWNER ou ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AiSettingsPatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/media:
    post:
      summary: Déposer une image (JPEG, PNG ou WebP)
      description: >
        Envoi multipart. Le type réel est déduit des octets, jamais du nom de
        fichier. Le média reste TEMPORARY tant qu'il n'est pas rattaché à une
        publication, puis il est déplacé vers la clé définitive.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                brandId: { type: string, format: uuid }
                purpose: { type: string, enum: [publication, avatar], default: publication }
      responses:
        '201': { $ref: '#/components/responses/Media' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/MediaTooLarge' }
        '415': { $ref: '#/components/responses/MediaTypeNotAllowed' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/media/{mediaId}:
    parameters:
      - { name: mediaId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Métadonnées et URL de lecture signée
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Media' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Supprimer un média non utilisé
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Média supprimé }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/publications:
    get:
      summary: Lister les publications accessibles
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: brandId, in: query, schema: { type: string, format: uuid } }
        - name: status
          in: query
          schema:
            type: string
            enum: [all, draft, scheduled, publishing, published, partially_published, failed, cancelled]
        - { name: provider, in: query, schema: { type: string, enum: [facebook, instagram] } }
        - { name: search, in: query, schema: { type: string, maxLength: 120 } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { $ref: '#/components/responses/PublicationList' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
    post:
      summary: Créer un brouillon (rôle COMMUNITY_MANAGER minimum)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PublicationCreate' } } }
      responses:
        '201': { $ref: '#/components/responses/Publication' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/publications/counts:
    get:
      summary: Compter les publications par statut
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: brandId, in: query, schema: { type: string, format: uuid } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /api/v1/publications/{publicationId}:
    parameters:
      - { name: publicationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Détail d'une publication
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Publication' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Modifier une publication non envoyée
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PublicationPatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Publication' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: Supprimer une publication (suppression logique)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Publication supprimée }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/publications/{publicationId}/schedule:
    parameters:
      - { name: publicationId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Planifier l'envoi
      description: >
        `scheduledAt` est un instant absolu (ISO 8601 avec décalage), stocké en
        UTC. `timezone` sert uniquement à réafficher l'heure locale voulue.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleRequest' } } }
      responses:
        '201': { $ref: '#/components/responses/Publication' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    patch:
      summary: Replanifier l'envoi
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/Publication' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
    delete:
      summary: Annuler la planification (retour au brouillon)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Publication' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/publications/{publicationId}/publish:
    parameters:
      - { name: publicationId, in: path, required: true, schema: { type: string, format: uuid } }
      - name: Idempotency-Key
        in: header
        required: false
        schema: { type: string, maxLength: 200 }
        description: Rejoue la réponse mémorisée si la même commande est renvoyée.
    post:
      summary: Demander un envoi immédiat
      description: >
        Réponse `202` : l'envoi est exécuté par le worker. La publication passe
        en `publishing`, puis en `published`, `partially_published` ou `failed`.
      security: [{ bearerAuth: [] }]
      responses:
        '202': { $ref: '#/components/responses/PublishAccepted' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/publications/{publicationId}/retry:
    parameters:
      - { name: publicationId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 200 } }
    post:
      summary: Relancer les réseaux en échec
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                provider: { type: string, enum: [facebook, instagram] }
      responses:
        '202': { $ref: '#/components/responses/PublishAccepted' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/calendar:
    get:
      summary: Publications planifiées ou publiées sur une période
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: brandId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/PublicationList' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/social-accounts:
    get:
      summary: Lister les comptes sociaux liés à une marque (Sprint 06)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: brandId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/social-accounts/{provider}/connect:
    parameters:
      - { name: provider, in: path, required: true, schema: { type: string, enum: [facebook, instagram] } }
    post:
      summary: Démarrer (ou relancer) une connexion OAuth Meta ; provider=instagram non supporté avant le Sprint 07
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brandId, mobileRedirectUri]
              properties:
                brandId: { type: string, format: uuid }
                mobileRedirectUri: { type: string, description: 'Deep link Expo, ex. hootly://oauth/callback' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/social-accounts/oauth/status:
    get:
      summary: Statut d’un state OAuth émis par /connect (PENDING/COMPLETED/EXPIRED)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/social-accounts/{socialAccountId}:
    parameters:
      - { name: socialAccountId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Déconnecter un compte social (révoque côté Social ; ADMIN ou plus)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Déconnecté }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/social-accounts/{socialAccountId}/sync:
    parameters:
      - { name: socialAccountId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Revalider la connexion auprès de Meta (pas de sync métriques avant le Sprint 12 — commentaires depuis le Sprint 08, voir /api/v1/comments/sync)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/comments:
    get:
      summary: Lister les commentaires d'une marque, paginé et filtré (Sprint 08, filtres IA au Sprint 09)
      description: >
        Les filtres d'analyse (sentiment/intention/priorité) portent sur la
        DERNIÈRE analyse du commentaire : un commentaire pas encore analysé
        n'en a aucune et sort donc des résultats dès qu'un de ces filtres est
        posé, plutôt que d'être rangé par défaut dans « neutre ».
        sort=priority classe HIGH d'abord, puis par date, les non analysés en
        dernier.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: brandId, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string, enum: [new, processed, ignored, escalated, all] } }
        - { name: network, in: query, schema: { type: string, enum: [facebook, instagram] } }
        - { name: publicationId, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string, maxLength: 120 } }
        - { name: sentiment, in: query, schema: { type: string, enum: [positive, neutral, negative, all] } }
        - { name: intent, in: query, schema: { type: string, enum: [question, info_request, complaint, claim, other, all] } }
        - { name: priority, in: query, schema: { type: string, enum: [low, medium, high, all] } }
        - { name: sort, in: query, schema: { type: string, enum: [recent, priority], default: recent } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/comments/counts:
    get:
      summary: Compteurs de commentaires par marque (Sprint 08, highPriority réel depuis le Sprint 09)
      description: >
        highPriority compte les commentaires ENCORE À TRAITER dont la dernière
        analyse est « high ». pendingAiResponses vaut toujours 0 : rien ne
        persiste encore de proposition de réponse (Sprint 10) — ici 0 signifie
        « aucune », jamais « inconnu ».
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: brandId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/comments/{commentId}:
    parameters:
      - { name: commentId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Détail d'un commentaire (Sprint 08)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        language: { type: string, enum: [fr, en, ar], default: fr }
        timezone: { type: string, example: Europe/Paris }
...
  responses:
...
    Session:
      description: Tokens rotatifs et utilisateur authentifié
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/SuccessEnvelope' }
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/SessionData' }
...
    ValidationFailed:
      description: Entrée invalide
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    Conflict:
      description: Ressource déjà existante ou état incompatible
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...